requirements.txt
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
scikit-image==0.22.0
matplotlib==3.8.2
